core: Move some PIN-related functions to apps.common.request_pin and implement verify_user_pin().
